create util function FindOperatorPosition for operators to locate proofs in resign and reshare
