+++ initiator.go
@@ -2,8 +2,6 @@ package spec
 
 import (
 	"fmt"
-
-	eth_crypto "github.com/ethereum/go-ethereum/crypto"
 )
 
 // RunDKG is called when an initiator wants to start a new DKG ceremony
@@ -82,29 +80,4 @@ func RunResign(signedResign *SignedResign) ([][]*Result, error) {
 		}
 	}
 	return results, nil
-}
-
-func GetBulkMessageHash(bulkMsg []SSZMarshaller) ([32]byte, error) {
-	hash := [32]byte{}
-	msgBytes := []byte{}
-	for _, resign := range bulkMsg {
-		resignBytes, err := resign.MarshalSSZ()
-		if err != nil {
-			return hash, err
-		}
-		msgBytes = append(msgBytes, resignBytes...)
-	}
-	copy(hash[:], eth_crypto.Keccak256(msgBytes))
-	return hash, nil
-}
-
-func GetReqIDFromMsg(instance SSZMarshaller) ([24]byte, error) {
-	// make a unique ID for each reshare using the instance hash
-	reqID := [24]byte{}
-	msgBytes, err := instance.MarshalSSZ()
-	if err != nil {
-		return reqID, err
-	}
-	copy(reqID[:], eth_crypto.Keccak256(msgBytes))
-	return reqID, nil
 }
